[CodecImage] Verify that CopyTexImage() is functionally a no-op
We are looking to eliminate StreamTextureSharedImageInterface
subclassing GLImage as part of eliminating GLImage as a public
interface. The blocker to doing so is that
SharedImageVideoSurfaceTexture passes a
StreamTextureSharedImageInterface instance to
AbstractTexture::BindStreamTextureImage(), which operates on the
passed-in object as a GLImage. That invocation will in turn result in
the StreamTextureSharedImageInterface instance's BindTexImage() or its
CopyTexImage() method being invoked.

StreamTextureSharedImageInterface has two implementations, and their
implementations of these methods are as follows:

- StreamTexture::BindTexImage(): no-op
- StreamTexture::CopyTexImage(): no-op
- CodecImage::BindTexImage(): no-op
- CodecImage::CopyTexImage(): renders to front buffer

Hence, we need to focus only on CodecImage::CopyTexImage(). Our
suspicion is that it is currently always the case that
CodecOutputBufferRenderer::RenderToTextureOwnerFrontBuffer() has already
been invoked at this point and has either (a) rendered to the front
buffer or (b) failed because it's in a non-renderable state, making
CodecImage::CopyTexImage() itself a no-op.

To verify this hypothesis, this CL adds a DumpWithoutCrashing() call
if CodecOutputBufferRenderer::RenderToTextureOwnerFrontBuffer() will
not be a no-op when invoked via CodecImage::CopyTexImage().

Bug: 1310020
Change-Id: I5b3b4f90e018183edab75e9b92c192f8be6d7962
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3908011
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: vikas soni <vikassoni@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049248}
NOKEYCHECK=True
GitOrigin-RevId: 0f61771750c4a95e03991766869ee30177221b99