Set packet pts from buffer
It is the case now that at least the vorbis decoder uses |frame->pts| in
order to determine the firstness of a frame, which means that the
timestamp we have in DecoderBuffer needs to be applied to the AVPacket
as well.

This is part of the ffmpeg roll.

Bug: 1366548
Change-Id: I826c6336d7a09f38efa2b8eb418c744fc7841460
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3911043
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1050100}
NOKEYCHECK=True
GitOrigin-RevId: 3ebc763b10099c875f2cc344791aca81f09c729e