VEAAdapter: Perform video frame processing only if required
VEAAdapter perform video frame processing (e.g. scaling, cropping and
format conversion) every input VideoFrame if VideoEncodeAccelerator
is configured I420+SHMEM. The processing is not required when the
frame is mappable, format is I420, the visible rectangle is the
resolution to be encoded and its origin is (0, 0). This CL changes
VEAAdapter to do the processing only if it is required.
The frame processing took 0.20ms (simple copy case I420->I420) in the
case of 1280x780 on volteer.

BUG=b:241506766
TEST=Camera source in codec-compare.glitch.me with disabling zero-copy video capture

Change-Id: Ie68bd7c397383fc26a7642f2668a745acb677efd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3907268
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1050500}
NOKEYCHECK=True
GitOrigin-RevId: a5f9814394ebd4016af256bd6a86e6ae1140255a