Send HEVC HDR display metadata to decoder.
HEVC HDR mastering displaying content volume information and content light level information may not be included in mdcv/clli box, but instead be present in prefix-sei if it is not Atmos. In that case, we rely on extra data after demuxing to parse corresponding prefix-sei and populate the HDR display metadata, and carry it over to decoder for delivering to display.

Bug: 1362411,1362288
Change-Id: I3fa419aea71216c17da5b3651e3d4af7ab8e220e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3916840
Commit-Queue: Jianlin Qiu <jianlin.qiu@intel.com>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052136}
NOKEYCHECK=True
GitOrigin-RevId: 85af8764718f33f0d221e96f31f5d993f364b4a2