[Cast Streaming] Remove animated_playout_delay
This patch removes the animated_playout_delay field
from the cast configuration, as it is only used as an initial
value and is unnecessary--especially since adaptive playout delay
is currently disabled.

When we re-enable adaptive playout delay, the max_playout_delay can be
a good starting value with the potential to lower the playout delay if
conditions allow.

Bug: 1363694
Change-Id: I4ddfa7addd9fd95ab28fd79843d0f9adae8217f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914834
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052227}
NOKEYCHECK=True
GitOrigin-RevId: cdec927fc2ac99132e17c8462eafe4a37e9b2697