VideoFrame::RequiresExternalSampler(): Fix DCHECK for Fuchsia.
This CL is a follow-up to CL:3908248. Fuchsia violates the DCHECK in
VideoFrame::RequiresExternalSampler() that asserts that the texture
target is GL_TEXTURE_EXTERNAL_OES if RequiresExternalSampler() returns
true. For Fuchsia, it looks like the target can be legitimately 0. The
FuchsiaVideoDecoder creates mailboxes out of GpuMemoryBuffers for the
output frames starting at [1]. Then a VideoFrame for that mailbox is
created starting at [2], and that's the VideoFrame that's sent out to
the client. The way the MailboxHolders that are wrapped in the
VideoFrame are created is by simply storing the Mailbox and a SyncToken,
and no target is specified (see [3]), so the `MailboxHolder` should
retain its default target of 0 [4].

This CL simply allows a target of 0 in addition to
GL_TEXTURE_EXTERNAL_OES.

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/fuchsia/video/fuchsia_video_decoder.cc;l=575-577;drc=8240038c4deccc46c5c7fed0679f7f7383d0c1a9
[2] https://source.chromium.org/chromium/chromium/src/+/main:media/fuchsia/video/fuchsia_video_decoder.cc;l=606-610;drc=8240038c4deccc46c5c7fed0679f7f7383d0c1a9
[3] https://source.chromium.org/chromium/chromium/src/+/main:media/fuchsia/video/fuchsia_video_decoder.cc;l=152-155;drc=8240038c4deccc46c5c7fed0679f7f7383d0c1a9
[4] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/common/mailbox_holder.cc;l=11;drc=05dfbc82b07e06f610b8f2fecaa4d272430cc451

Bug: b:185790070, b:195769334, 1371201
Test: None
Change-Id: I7d922f024b620ccbf5c2d838195970f5cc83df00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3926381
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055053}
NOKEYCHECK=True
GitOrigin-RevId: e7bd75d6176f00db0ff4f14cc1c28a0a629faba5