Apply volume smoothing to audio start
This CL adds a small 5ms linear fade-in when first playing audio. This
removes a small pop, from the discontinuity in starting/resuming audio
play back.

Adding a small fade-out when pausing audio was attempted, but dropped.
By the time we pause, there is no new data coming in, and we cannot
fade it out. Replaying the last buffer did not help, and implementing
a pause-lookahead would be complex, and might introduce latency to the
pipeline.

Bug: 1337358
Change-Id: I904494e1dfa04f67944663db9e04e4d455f28fc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3949746
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1060665}
NOKEYCHECK=True
GitOrigin-RevId: ed3d24f3d66a5050459f3a215d6ae5394fb9c4d1