media: Add Callback Timeout Helpers
This CL adds helpers to catch callback timeouts. This can be used for
logging/metrics to identify potential callback hangs, or for handling
long timeouts to improve user experience.

See BUG for the link to the design doc.

The impl is similar to mojo/public/cpp/bindings/callback_helpers.h.
The new class is in media/base since it's useful in non-mojo cases as
well.

Bug: 1374532, b/253527777
Test: Added unittests
Change-Id: Ie56156d6af3a9abb4bc5c5687ac86cba4c4075c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3953267
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1061244}
NOKEYCHECK=True
GitOrigin-RevId: eae81e605ab619bf1dd0ced1ef70dc9e9b82aaad