Linux/VA-API: Set 'RGBX_8888' usage as 'GPU_READ' for VDA
This CL fixes VDA failure with Chromium minigbm on Linux.
The 'VaapiVideoDecodeAccelerator' requires the VPP to download the
decoded frame from the internal surface to the allocated native pixmap.
'SCANOUT_VDA_WRITE' is used for 'YUV_420_BIPLANAR' on ChromeOS; For
Linux, the usage is set to 'GPU_READ' for 'RGBX_8888'.

Bug: 1338827
Change-Id: I0d623b27503e64837a82a254a88a6c58e693e568
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757114
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1061367}
NOKEYCHECK=True
GitOrigin-RevId: 9bc26a79377b2b07e626beaf0dd2532dad1231bd