Enable Lacros Aura Capture
The main reason this was disabled when initially checked in was due to
some issues seen with delegated compositing during initial
implementation a few months ago. However, I've verified on a dev channel
build that this now seems to be working as expected.

Fixed: 1273189
Change-Id: I51eb30a6779f1358274080fc55d30e8dda44d47b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3972158
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1062330}
NOKEYCHECK=True
GitOrigin-RevId: 7453fa858609832b17196df11331972855da59bc