[video capture] Disable GMB capture on windows for d3d9 gpu mode
GMBs on windows are backed by d3d11 textures which are not supported in
the d3d9 mode.

This CL makes to fixes:

1) When starting the video capture utility process, if the GPU is in the
d3d9 mode, GMB switch is not appended now.

2) If the GPU process restarts and loses d3d11 support while the capture
is ongoing, theres's now a fallback to premapped frames in VideoCaptureImpl if the d3d shared images are not supported.
Gpu::Capabilities are now exposed by GpuVideoAcceleratorFactories and VideoCaptureImpl tests are updated.

Bug: 1366454
Change-Id: Ic3cf3b2d5e944929fcb54372f739c0567258b1c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3964998
Commit-Queue: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1062773}
NOKEYCHECK=True
GitOrigin-RevId: cd9de4d211ae882a15d8ea43a957cb2f0e2a25e0