NullAudioSink: Use SEQUENCE_CHECKER for control signals
Currently NullAudioSink checks that the Stop(), Start(), Play() and
Pause() methods are called on the same sequence that is used
internally by the FakeAudioWorker.

However, after some discussion it seems like we only need to check
that these functions are called in a valid sequence.  It can be a
different sequence than the FakeAudioWorker sequence.

This CL uses base/sequence_checker to do so.

Relaxing the checks in this way should allow NullAudioSink to be used
directly for WebAudio silent sink processing.

Reference: https://crrev.com/c/3924220

Change-Id: I3755e0947c046cd8e9f84badbe45fe47d1d4a958
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3975868
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1062978}
NOKEYCHECK=True
GitOrigin-RevId: f1f444a8aff122848f9cab4569e870447e7ce3d9