media/gpu/chromeos: Add DmabufVideoFramePool::GetGpuBufferLayout()
This CL adds said method and the needed implementation, to avoid
having to reinitialize the VideoFrame pool in V4L2VideoDecoder just
to know the modifier.

This is needed for https://crrev.com/c/3981035 which moves the
calculation of the amount of VideoFrames needed by the Renderer
pipeline from the (VAAPI,V4L2*)VideoDecoder to its parent
VideoDecoderPipeline.

Bug: b:250187354, b:254657177
Change-Id: I6f68bf2fcf4aab9078aca593c8ffc6c859d3d2fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3991797
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1066024}
NOKEYCHECK=True
GitOrigin-RevId: f7d7aa0bb512e2693b9313c600b4755a9766703b