media/gpu: Re-enable V4L2 decoders for standard Linux if built
Previously the code that interfaces with V4L2 decoder classes were
only guarded by the "USE_V4L2_CODEC" build flag. This implied the
that as long as the build flag was set, it could be built and run
on standard Linux.

Recently code was added to guard it to only use it for ChromeOS
Ash (and not for LaCrOS), which inadvertently disabled the code
for Linux as well.

As we want to develop the V4L2 decoders on standard Linux in the
future, fix up the build flag guards so that it works when built
for Linux with "use_v4l2_codec=true" manually specified.

      Build for standard Linux with "use_v4l2_codec=true" GN arg.
      VDA test should be able to find V4L2 decoder and try to use it,
      but likely fail (due to other reasons)

Bug: None
Test: Extract newer "linux-libc-dev" >= 5.17 package to sysroot, then
Change-Id: Ie670e0a25ec6567dafc60c080a892905e2384d23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3996068
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1067480}
NOKEYCHECK=True
GitOrigin-RevId: 149471b4a55327f34b1130a71bb85aff4d35d487