HLS: Implement alternative audio renditions
This CL implements the `AudioRendition` and `AudioRenditionGroup`
classes in order to support alternative audio renditions in multivariant
playlists. This is particularly important because it's common for HLS
presentations to not include audio in the primary rendition of each
variant, relying on the client to select an appropriate audio rendition
instead.

Bug: 1266991
Change-Id: I2378938b3c7c9a5c78fec7e63f6ea3cec1700fb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3858521
Reviewed-by: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1067782}
NOKEYCHECK=True
GitOrigin-RevId: 499dc5eb3c98c3f39460a2d533fa5b815163cdc9