sandbox: Introduce hardware video decoding BPF policy.
This CL creates a sandbox BPF policy for hardware video decoding (in the
context of out-of-process video decoding). For the purposes of this CL,
the BPF policy is the same as the GPU process BPF policy. In follow-up
CLs, the BPF policy will be refined for the different possibilities.
These possibilities are encoded in ComputePolicyType() which is
introduced in this CL. Follow-up CLs will use the return value of this
method to configure both the pre-sandbox hook and the BPF policy itself.

Bug: b:210759684
Test: None
Change-Id: Ie7e8db75ae7c1529030d53dc876426f94dbd040b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3997594
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1069372}
NOKEYCHECK=True
GitOrigin-RevId: cd4605bc80818c1fca8a96e73d34d7b6d8fa9d74