Fix viz::ResourceFormat for display of software RGB VideoFrames.
We were always using PaintCanvasVideoRenderer for processing RGB
VideoFrames for upload, but were telling the resource the original
frame's format instead of the one that PCVR was converting to.

Now we'll only use PCVR if we can't upload directly from the RGB
resource. When we do use PCVR we'll also now tag the resource with
the format that PCVR outputs.

We can avoid using PCVR when the frame format is close enough to
the output resource format (RGBA/RBGX into RGBA, etc).

This also fixes a few other bugs with the RGB conversion path:
- Y16 or YUV w/ disable_one_component_textures had the same issue.
- Uploads were done at 0, 0 instead of visible_rect origin.

Fixed: 1380315

Change-Id: Ic66865bd54ead32dff0e0a100113a50c0ae75be9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4018480
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1070039}
NOKEYCHECK=True
GitOrigin-RevId: 2296f19c8d06f3be15f39bc84bd89d6f2a696935