Video: Filter unsupported d3d11 profiles for dxva decoder
If d3d11 video decoder is disabled by gpu workaround, when we query the
supported decoder profile, it will fallback to query dxva supported
profiles, this is correct. But dxva profiles uses all d3d11 profiles
as its supported profiles, and never filter those unsupported profiles
in the `kSupportedProfiles` list, this is wrong.

Add a logic to ignore profiles that are not in the dxva supported
profile list.

Bug: 1384481
Change-Id: I4a57e5cf9402af1ba09c763fea43bfe9d834246e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4028257
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: 朱思达 <zhusida@bytedance.com>
Cr-Commit-Position: refs/heads/main@{#1072015}
NOKEYCHECK=True
GitOrigin-RevId: b85194db3f277470194b6c2554a12eb1603a9b2e