[media] Eliminate GLImage knowledge from VaapiPictureNativePixmap
We are trying to eliminate GLImage. VaapiPictureNativePixmap has no need
to know about GLImage: it currently holds a |gl_image_| ivar simply for
convenience of its subclasses. This CL moves that ivar into the concrete
subclasses that need it and eliminates VaapiPictureNativePixmap knowing
about GLImage. This helps to clarify the actual GLImage elimination work
that is necessary in this space.

Bug: 1366367
Change-Id: I9df8b3199e084350598cfba106dd082634506405
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4025005
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1072105}
NOKEYCHECK=True
GitOrigin-RevId: a85ebcdf388cfd20c52a189a5924cf80d19e79ea