Fix access of uninitialized variable in MediaMetricsProvider
In some cases, the destructor for `MediaMetricsProvider` could be called
without a prior call to `MediaMetricsProvider::Initialize`, resulting in
an uninitialized read to `MediaMetricsProvider::media_stream_type_` in
the destructor. Since there isn't a sensible default value to assign to
that member, this CL wraps it in an optional to prevent access prior to
initialization and prevents unchecked access in the destructor. This
implementation was based off the discussion in
https://chromium-review.googlesource.com/c/chromium/src/+/2964116

Bug: 1220321
Change-Id: I17382cd191da91d66d3ff1b2f95fc210106739d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4028793
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1072531}
NOKEYCHECK=True
GitOrigin-RevId: 2b3ab1bd80da862c2c2d9656f89d7503c6bd1e78