OOP-VD: Re-use StableCdmContextImpl across Initialize()s.
CL:4026445 revealed that video.PlayDRM.cencv3_h264_cbc_then_ctr_oopvd
fails. The reason is that MojoVideoDecoderService has logic to ensure
that the CDM doesn't change across Initialize() calls [1]. Today, the
StableVideoDecoderService (the VideoDecoder entry point in the video
decoder utility process) unregisters the CDM context on Initialize() if
one was already registered by a previous Initialize() call [2]. That
means that a new CDM ID is assigned at each Initialize() call that sets
up encryption. This triggers the logic in the MojoVideoDecoderService
[1].

This CL addresses that by re-using the same StableCdmContextImpl (GPU
process) and corresponding RemoteCdmContext (utility process) across
initializations. It's then only necessary to provide a
pending_remote<StableCdmContext> to the utility process the first time
the decoder is Initialize()d for protected content.

Versioning considerations: the out-of-process video decoding feature has
not been enabled by default yet and it's not turned on in an experiment.
Therefore, it's not important to maintain the previous behavior yet.
However, for the sake of completeness, here are the two potentially
problematic scenarios:

- A newer client running with an older service: the service is expected
  to accept only the first Initialize() call that sets up encryption and
  reject subsequent ones because the client won't supply a bound
  pending_remote for the latter thus triggering [3].

- An older client running with a newer service: the service is expected
  to maintain the connection to the first StableCdmContext alive across
  Initialize() calls and drop the subsequent
  pending_remote<StableCdmContext> supplied by the older client. This
  should be functionally fine.

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/mojo/services/mojo_video_decoder_service.cc;l=231-246;drc=d387f092f963cdceb9568d3a589686f7cf0e5986
[2] https://source.chromium.org/chromium/chromium/src/+/main:media/mojo/services/stable_video_decoder_service.cc;l=100-105;drc=d387f092f963cdceb9568d3a589686f7cf0e5986
[3] https://source.chromium.org/chromium/chromium/src/+/main:media/mojo/services/stable_video_decoder_service.cc;l=110-116;drc=d387f092f963cdceb9568d3a589686f7cf0e5986

Bug: b:194429120, b:195769334
Test: tast run ${IP} video.PlayDRM.*_oopvd on volteer (CL:4026445)
Change-Id: Id909f44db8a6f975a9a27ed153bb40f5a918d457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4022802
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1073054}
NOKEYCHECK=True
GitOrigin-RevId: b57974b634738e9dfb96e5154101d713b3b62ee0