Fix behavior of DefaultEnumValue vs kOk
There was previously some confusion about the intention of a "Default
Value" for status code enums. At some point we incorrectly changed the
implementation so that only status enums which declared a literal value
with the name "kOk" could ever be considered as an "ok status", which
caused significant unknown behavior with our long-standing
PipelineStatus implementation.

This change makes it so that the code now lines up with the
documentation, which had been correct all along. It also is able to
clean up some of the enable_if mess along the way :)

The general gist now that if a traits struct declares a static constexpr
method |DefaultEnumValue|, then the return from that function is the
"OK" value for the status. Additionally, if this method is omitted, BUT
the enum has a literal value named "kOk", then this is used as the "OK"
value for the status. It is Acceptable for a status to NOT have an "OK"
value, however, certain operations wont work, such as
TypedStatus<T>::Or<Other>::code(). Finally, if BOTH the DefaultEnumValue
method and kOk are present, then kOk MUST be the return from
DefaultEnumValue.

Bug: 1384672
Change-Id: Ibad06c53c902a225369ba75d04a4109d764c27e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4029383
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074303}
NOKEYCHECK=True
GitOrigin-RevId: 72b81e9005ace1c46d3b4b76af0f05a9f347e945