[GPU] Move GLImagePbuffer into //gpu/command_buffer/service/shared_image
This CL is a no-op change that moves GLImagePbuffer from being internal
to DXVAPictureBuffer to living in
//gpu/command_buffer/service/shared_image. This will allow us to narrow
GLImageBacking to taking in a GLImagePbuffer and becoming just
GLImagePbufferBacking in a followup change.

Note that we are intentionally not generalizing or renaming
GLImagePbuffer as we are looking to strip down and eventually remove
this codepath (and GLImageBacking) altogether, and so we want to keep
it completely restricted to its current use case.

Bug: 1384438
Change-Id: Ia031405a60b1e3a276b2d68cd4de88f49356af6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4036568
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074501}
NOKEYCHECK=True
GitOrigin-RevId: d1218e05972852ac27f95f962a03966b05d87ac3