[GLImageBacking] Hardcode to usage with GLImagePbuffer
GLImageBacking takes in a GLImage instance. In practice this is always
a GLImagePbuffer:
- The only way to create GLImageBacking is via CreateFromGLTexture()
- The only callsite of GLImageBacking::CreateFromGLTexture is in
  DXVAVideoDecodeAccelerator
- That class only calls the method if the GLImage instance that it
  obtains from DXVAPictureBuffer is *not* of type GLImageDXGI
- DXVAPictureBuffer creates only two types of GLImages:
  - GLImageDXGI (or CopyingGLImageDXGI, which inherits the same
    GLImage type)
  - GLImagePbuffer

This CL hardcodes GLImageBacking to take in a GLImagePbuffer, having
DXVAPicture safely downcast the GLImage instance that it passes in
(which, as above, we know is actually a GLImagePbuffer instance). We
also rename GLImageBacking to GLImagePbufferBacking.

This change will allow us to drastically simplify
GLImagePbufferBacking for the fact that GLImagePbuffer is essentially a
dummy class.

Bug: 1384438
Change-Id: Ibca0b60185810a92e19c8d9ceaee5121038f670e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4030407
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074525}
NOKEYCHECK=True
GitOrigin-RevId: fc1b380152425446941546019715675c2afb0890